Update: Moved images dir to assets dir
