+++ src/components/HomePage/WelcomeScreen/WelcomeScreen.tsx
@@ -1,8 +1,7 @@
 import React, { EffectCallback, FunctionComponent, ReactElement, useEffect } from 'react'
 
+import { deejay250x250 } from '@/assets/images'
 import ViewMyWorkButton from '@/components/HomePage/ViewMyWorkButton'
-
-import { deejay250x250 } from '@/common/images'
 
 import { executeTitleTypeAnimation } from './helpers'
 import './styles.css'
